Fixes for scenes.
Destroy shader variable group and material descriptions after the other
resources. This avoids invalid memory access if destroyed before materials.

Fixed normals for OBJ files.

+++ modules/Scene/src/SceneResources.c
@@ -370,13 +370,30 @@ void dsSceneResources_freeRef(dsSceneResources* resources)
 	if (DS_ATOMIC_FETCH_ADD32(&resources->refCount, -1) != 1)
 		return;
 
+	// First pass: free everything but descriptions.
 	for (dsListNode* node = resources->resources->list.head; node; node = node->next)
 	{
 		ResourceNode* resourceNode = (ResourceNode*)node;
 		if (resourceNode->type == dsSceneResourceType_SceneNode)
 			dsSceneNode_freeRef((dsSceneNode*)resourceNode->resource);
-		else if (resourceNode->owned)
+		else if (resourceNode->owned &&
+			resourceNode->type != dsSceneResourceType_ShaderVariableGroupDesc &&
+			resourceNode->type != dsSceneResourceType_MaterialDesc)
+		{
 			destroyResource(resourceNode->type, resourceNode->resource);
+		}
+	}
+
+	// Second pass: destroy descriptions.
+	for (dsListNode* node = resources->resources->list.head; node; node = node->next)
+	{
+		ResourceNode* resourceNode = (ResourceNode*)node;
+		if (resourceNode->owned &&
+			(resourceNode->type == dsSceneResourceType_ShaderVariableGroupDesc ||
+				resourceNode->type == dsSceneResourceType_MaterialDesc))
+		{
+			destroyResource(resourceNode->type, resourceNode->resource);
+		}
 	}
 
 	if (resources->allocator)
